Refactor: Extract getError function into util folder
