+++ react/store/createReturnRequest/ReturnDetailsContainer.tsx
@@ -2,14 +2,12 @@ import React, { useEffect, useState } from 'react'
 import { useRuntime } from 'vtex.render-runtime'
 import { PageHeader, PageBlock } from 'vtex.styleguide'
 import type { RouteComponentProps } from 'react-router'
-import { defineMessages, FormattedMessage } from 'react-intl'
+import { FormattedMessage } from 'react-intl'
 import { useQuery } from 'react-apollo'
 import type {
   OrderToReturnSummary,
-  OrderToReturnValidation,
   QueryOrderToReturnSummaryArgs,
 } from 'vtex.return-app'
-import type { ApolloError } from 'apollo-client'
 
 import { StoreSettingsPovider } from '../provider/StoreSettingsProvider'
 import { OrderToReturnProvider } from '../provider/OrderToReturnProvider'
@@ -17,62 +15,11 @@ import { ReturnDetails } from './components/ReturnDetails'
 import { ConfirmAndSubmit } from './components/ConfirmAndSubmit'
 import { useReturnRequest } from '../hooks/useReturnRequest'
 import ORDER_TO_RETURN_SUMMARY from './graphql/getOrderToReturnSummary.gql'
-import { ORDER_TO_RETURN_VALIDATON } from '../utils/constants'
 import { formatItemsToReturn } from '../utils/formatItemsToReturn'
 import { setInitialPickupAddress } from '../utils/setInitialPickupAddress'
+import { getErrorCode, errorMessages } from '../utils/getErrorCode'
 
 export type Page = 'form-details' | 'submit-form'
-
-type CodeError =
-  | 'UNKNOWN_ERROR'
-  | 'E_HTTP_404'
-  | 'FORBIDDEN'
-  | OrderToReturnValidation
-
-const { ORDER_NOT_INVOICED, OUT_OF_MAX_DAYS } = ORDER_TO_RETURN_VALIDATON
-
-const getErrorCode = (error: ApolloError): CodeError => {
-  const { graphQLErrors } = error
-
-  if (!graphQLErrors.length) {
-    return 'UNKNOWN_ERROR'
-  }
-
-  const [{ extensions }] = graphQLErrors
-
-  const { code } = extensions?.exception ?? {}
-
-  const knownErrors = [
-    ORDER_NOT_INVOICED,
-    OUT_OF_MAX_DAYS,
-    // order not found
-    'E_HTTP_404',
-    // userId on session doesn't match with userId on order profile
-    'FORBIDDEN',
-  ] as const
-
-  const knownError = knownErrors.find((errorCode) => errorCode === code)
-
-  return knownError ?? 'UNKNOWN_ERROR'
-}
-
-const errorMessages = defineMessages({
-  [ORDER_NOT_INVOICED]: {
-    id: 'store/return-app.return-order-details.error.order-not-invoiced',
-  },
-  [OUT_OF_MAX_DAYS]: {
-    id: 'store/return-app.return-order-details.error.out-of-max-days',
-  },
-  E_HTTP_404: {
-    id: 'store/return-app.return-order-details.error.order-not-found',
-  },
-  FORBIDDEN: {
-    id: 'store/return-app.return-order-details.error.forbidden',
-  },
-  UNKNOWN_ERROR: {
-    id: 'store/return-app.return-order-details.error.unknown',
-  },
-})
 
 type RouteProps = RouteComponentProps<{ orderId: string }>
